src/macro/bdvfl/requirements.txt
torch
torchvision
pysha3
requests
web3==5.15.0